requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
httpx
python-dotenv
requests
flask
